split up entry point generation/deletion
