+++ package.json
@@ -10,7 +10,7 @@
     "dev:helpers": "cd packages/convex-helpers && chokidar '*.ts' 'server/**/*.ts' 'react/**/*.ts*' 'tsconfig*.json' 'package.json' -c 'npm run build' --initial",
     "predev": "cd packages/convex-helpers && npm run build",
     "lint": "tsc --project tsconfig.test.json",
-    "build": "cd packages/convex-helpers && node generate-exports.mjs && tsc",
+    "build": "node generate-exports.mjs && cd packages/convex-helpers && tsc",
     "clean": "rm -rf packages/convex-helpers/dist",
     "test": "vitest run",
     "test:watch": "vitest",
